+++ KarteInAppMessaging/InAppMessaging.swift
@@ -225,10 +225,10 @@ extension InAppMessaging {
         }
 
         if window.isKind(of: IAMWindow.self), let delegate = delegate {
-            if let scene = window.windowScene {
+            let selector = #selector(InAppMessagingDelegate.inAppMessagingWindowIsPresented(_:onScene:))
+            if (delegate as AnyObject).responds(to: selector), let scene = window.windowScene {
                 delegate.inAppMessagingWindowIsPresented?(self, onScene: scene)
             } else {
-                Logger.warn(tag: .inAppMessaging, message: "WindowScene not found. Skipped calling inAppMessagingWindowIsPresented(_:onScene:)")
                 delegate.inAppMessagingWindowIsPresented?(self)
             }
         }
@@ -253,10 +253,10 @@ extension InAppMessaging {
         }
 
         if window.isKind(of: IAMWindow.self), let delegate = delegate {
-            if let scene = window.windowScene {
+            let selector = #selector(InAppMessagingDelegate.inAppMessagingWindowIsDismissed(_:onScene:))
+            if (delegate as AnyObject).responds(to: selector), let scene = window.windowScene {
                 delegate.inAppMessagingWindowIsDismissed?(self, onScene: scene)
             } else {
-                Logger.warn(tag: .inAppMessaging, message: "WindowScene not found. Skipped calling inAppMessagingWindowIsDismissed(_:onScene:)")
                 delegate.inAppMessagingWindowIsDismissed?(self)
             }
         }